requirements.txt
Django==4.0.4
djangorestframework==3.13.0
pyyaml
requests
django-cors-headers
djoser
djangorestframework-simplejwt